Move `stem` from preprocessing.py to nlp.py
